+++ ProgramFunctions/Update.fnc.php
@@ -472,7 +472,7 @@ function _update83()
  * Update to version 8.4
  *
  * 1. GRADEBOOK_GRADES table: Change comment column type to text.
- * 2. ACCOUNTING_INCOMES table: Add FILE_ATTACHED column.
+ * 2. accounting_incomes table: Add FILE_ATTACHED column.
  *
  * Local function
  *
@@ -495,7 +495,7 @@ function _update84()
 		ALTER COLUMN comment TYPE text;" );
 
 	/**
-	 * 2. ACCOUNTING_INCOMES table: Add FILE_ATTACHED column.
+	 * 2. accounting_incomes table: Add FILE_ATTACHED column.
 	 */
 	$file_attached_column_exists = DBGetOne( "SELECT 1 FROM pg_attribute
 		WHERE attrelid=(SELECT oid FROM pg_class WHERE relname='accounting_incomes')